+++ CHANGELOG.md
@@ -3,3 +3,4 @@ dev (1.4.0)
 
 * Remove unused Elasticsearch backup code (#8076)
 * Improve performance of indexing AIP by saving uncompressed METS (#7424)
+* Index MODS identifiers in Elasticsearch (#7424)